Use Alert mixin for login errors
