+++ src/comp/oauth/Login.js
@@ -1,0 +1,23 @@
+import Alert from "../Alert.js"
+
+export default {
+    ready() {
+        let promise = window.api.login(this.$route.query["access_token"])
+        promise.then(
+            (account) => {
+                this.account = account
+                this.$router.go({path: "/"})
+            },
+            (error) => {
+                this.reportError(error)
+                console.log(error)
+            }
+        )
+    },
+
+    mixins: [ Alert ],
+
+    props: {
+        account: { twoWay: true, required: true }
+    }
+}
